Gameplay implemented
- some small bug fixes

This is my last commit when it comes to recrutment work. It was quite a hustle. Got realy tired lately, when deving gameplay stuff and now it paid dividence in lots of bugs. There is still a bug that will insta kill enemies and spawners gonna get glithed, but since I implemented just simple pool and had not enought time to finish it I leav this bug there. Thank you test driven development for today.